Sort translations for a function by TransID
Summary:Sorting translations by bytecode order was kind of useful when we were compiling
tracelets, but that's not the common case anymore.  And I've found myself a few
times wanting to look at the translations for a function in the order they were
created, so sort translations by that instead.

Reviewed By: paulbiss

Differential Revision: D3181050

fb-gh-sync-id: 2e87a897fb2e728ced913066fa660a4c329f132a
fbshipit-source-id: 2e87a897fb2e728ced913066fa660a4c329f132a