Handle inheritance when looking up method declaration in IdentifySymbolService
Summary:Data in named AST is a file-level information that doesn't include inheritance, so looking up declaration of Foo::bar() was not working when bar() was not defined directly in Foo.

According to reports (#10838231) this supposedly used to work in Nuclide, but it's not clear to me how. The stuff modified in this diff is only used in a call we yet have to deploy, and current version (the "server only" one, without hh_ide.js) is using --identify-function and passing the resulting string (in this case: "Foo::bar") to search, which would not find it, since search is also a textual and inheritance unaware.

We made no functionality changes to --identify-function and --search recently, so I feel that the version that was using hh_ide.js was doing some additional heuristic for looking up names that got ripped off, but the code looks like sphagetti to me and I can't tell.  EDIT: I did tell, see comment: D3170873#inline-21301956

Anyway, next week we plan to switch the whole thing to --ide-get-definition call, so this diff fixes the issue there.

Relevant changes are in identifySymbolService.ml, the rest is threading of tcopt through everything which is necessary to access typing heap in presence of lazy decl.

Reviewed By: alexchow, ericlippert

Differential Revision: D3182588

fb-gh-sync-id: 61e581fbbae2c5f0bbe3932e2da79547e3dfc942
fbshipit-source-id: 61e581fbbae2c5f0bbe3932e2da79547e3dfc942