Add --ide-get-definition option
Summary:This uses the same RPC call as --identify-function and --get-method-name, but displays (even) more information - including the definition position if available, which enables efficient implementation of hyperclick (go to definition).

I'm not exposing it in --help to be able to iterate on it and possibly introduce breaking changes. Nuclide is fine with us being backwards compatible only for 2-3 versions, so if needed we can introduce ide-get-definition2 with breaking changes, and later on rename it back to ide-get-definition, and if the API matures we can document it.

peterhal Please take a look at the test plan for example session with this mode and let me know if you would like to make any changes in the API.

Reviewed By: int3

Differential Revision: D3169799

fb-gh-sync-id: 08e34b7cd812225ad67a0601b33258ba5934e5b6
fbshipit-source-id: 08e34b7cd812225ad67a0601b33258ba5934e5b6