Don't allow refs into containers during construction via arrays
Summary:During the conversion from IDL, an optimization was added to the container
classes for construction via arrays. If the array met the criteria for the
specific container, the container will simply store the entire array as its own
rather than copying the array's contents into a private copy. Unfortunately,
this optimization isn't completely sound as the source array can contain
references and containers can't contain references. As a result, if the source
array does contain a reference, invariants will be broken and cause incorrect
behavior afterwards.

The solution is to scan the array for references when deciding if the source
array meets the criteria for the optimization. This is still more efficient than
doing an element-wise insertion, but does make this optimization less useful. If
the source array does contain a reference, the optimization won't fire and we'll
resort to element-wise insertion, which will properly unbox the reference before
inserting it.

Also add a test-case which exercises this corner case.

Reviewed By: markw65

Differential Revision: D3189334

fb-gh-sync-id: e4d38092ad23d477b12279427484683f80f47339
fbshipit-source-id: e4d38092ad23d477b12279427484683f80f47339