Improve strcoll test portability
Summary:Accordingly to the man page [1] and php documentarion [2], the
strcoll function does not specify the return value for strcoll
when the operands are not equal. It just specifies if the return
value must be greater, or lower than zero in case the operands are
different.

For ppc64, when running the following chunk of code in a cpp
program (i.e. using string.h's strcoll function, like hhvm does),
I get "1" as result:

  setlocale(LC_COLLATE,"C");
  printf("%d\n", strcoll("a","A"));

However on x86 I get "32".

The results are different, but according to [1] and [2] they are
not wrong. The zend test [3] is done in a portable way.

[1] http://linux.die.net/man/3/strcoll
[2] http://php.net/manual/en/function.strcoll.php
[3] hphp/test/zend/good/ext/standard/tests/strings/strcoll.php
Closes https://github.com/facebook/hhvm/pull/6981

Reviewed By: Orvid

Differential Revision: D3169829

fb-gh-sync-id: 9d5fb792176d6a50c7ad3ab90035f1fda2493826
fbshipit-source-id: 9d5fb792176d6a50c7ad3ab90035f1fda2493826