Delay the effective start of profiling counters
Summary:This diff delays, by a given number of requests, when profiling counters effectively start to count.  This allows more profile translations to be produced before the counters effectively start, which can both (a) improve the accuracy of the counters and (b) allow for more portions of a function and different combinations of types to be seen before retranslating the function in optimized mode.  Both of these benefits come from avoiding issues caused by contention on the write lease when server is starting up and JITing heavily. For example, even if a portion of the function is hot, it may not be JITed soon because the threads trying to JIT it failed to get the write lease. As a result, such portion of the function may end up being JITed in Live mode and placed far away from the main (optimized) translation for that function.  The effect of contention on the write lease on the accuracy of the profile counters is similar: a profile translation may artificially end up with much higher profile count than others that are just as hot simply because it succeeded to be JITed sooner. This can distort the relative profile counts for profile translations of the same function, which guides some optimization passes (e.g. code layout).

The counting (decrement) of profiling counters needed to be changed to use the lock prefix to enforce atomic execution, so that we can make sure that they're properly reset.

This diff also fixes a pre-existing divide-by-zero bug when calculating IRGS.profFactor for inlined calls that got exposed.

Reviewed By: swtaarrs

Differential Revision: D3188879

fb-gh-sync-id: 84aada96e438c1042e425a155fbf470056ca4815
fbshipit-source-id: 84aada96e438c1042e425a155fbf470056ca4815