Simplify stack high water tracking
Summary:Instead of separately tracking the max depth of the eval stack, and
the max depth of the fpi stack, just track the max of the sum. This
results in a better bound. Also, use the FuncEmitter's maxStackCells
field to track the max depth, to make communication between the hhas
and php emitters more straightforward.

I did this as part of the inline assembly diff, but it got big enough
to be worth breaking out for separate review.

Reviewed By: mxw

Differential Revision: D3193045

fb-gh-sync-id: a98ab7f472e9b98adeef65f496c72b969a436b27
fbshipit-source-id: a98ab7f472e9b98adeef65f496c72b969a436b27