Log parse errors during repo building
Summary:Add an error message during hhvm repo building for all encountered syntax errors.
This is only informative and compilation proceeds as usual, but this should help to identify problems as soon as possible.
This should also provide an alternative to hhvm -l syntax check - see #6979 for issues with it.
Closes https://github.com/facebook/hhvm/pull/6988

Reviewed By: Orvid

Differential Revision: D3176516

fb-gh-sync-id: ad657996cb5eeeb2540a2ab9032b79b7f7c4e8ff
fbshipit-source-id: ad657996cb5eeeb2540a2ab9032b79b7f7c4e8ff