Initial data structures for local finding service
Summary:My re-implementation of the local finder service is done, but the diff is big enough that I want to break it up into many steps.

The basic plan here is to pass an accumulator along during the walk of an unnamed AST. The accumulator will collect information about each logically distinct local in the file, and when we're done, we can ask the accumulator for the positions of the local we're interested in.

These data structures will be used to form the accumulator, which will be presented in the following diff. Everything is purely functional.

The fundamental problem is that we need to deal with this situation:

     function test1($x, $y) {
       $f = function ($z) use($x) {
         $y = 100;
         return $x + $y + $z;
       };
       return $f(10);
     }

Inside test1 but outside of the anonymous function, $x, $y and $f are in scope.

Inside the anonymous function, $x, $y and $z are in scope, but $y is a distinct local from the formal parameter, and $f is not in scope. However $x is the same as the formal parameter.

Perhaps now it is clear why we need a stack of stacks of maps, but in case it is not, the structure of the scope chains needs to come out like this during the analysis of the body of the anonymous function:

     [
       [
         [ $y -> 5; $z -> 4 ]; [ $x -> 1 ]
       ];
       [
         [ $x -> 1; $y -> $2; $f -> 3 ]
       ]
     ]

When doing a name lookup, only the head scope chain is in effect, so a lookup of $f would fail to find an ident. (Code for that operation will be in the next diff.)

The careful reader may have noticed that there are two scopes in the scope chain for the anonymous function. In the corner case (almost certainly an error!) of "function($x) use($x)", the formal parameter wins. The easiest way to deal with this situation was to put the use clause scope deeper in the scope chain than the formal parameter scope.

Reviewed By: dabek

Differential Revision: D3176022

fb-gh-sync-id: 85a84a8057765c058ed2c9685352b1a0bbde89bf
fbshipit-source-id: 85a84a8057765c058ed2c9685352b1a0bbde89bf