Stop constraining pseudo-main guards eagerly.
Pre-constraining pseudo-main guards doesn't allow the guard relaxation pass to
take effect. The result is that a lot of unnecessary type guards are left in
pseudo-main. This diff stops constraining the guards eagerly and saves type
information for unconstrained locals into RegionDesc as "type predictions". This
decreases the total number of executed type guards for an oss-performance
WordPress run by 40%.