Track ctx type in FPIInfo
Summary:Since least_common_ancestor only tracks through passthrough instructions it
should be safe to assert that the type we get at the inlined call site is the
union of the merged types. Track this separately in the FPIInfo.

Also add an assert that the type we have is at least TCtx.

Reviewed By: ottoni

Differential Revision: D3204408

fb-gh-sync-id: 0883c6380e0900380ef4457dfa0d367b1bd0deb9
fbshipit-source-id: 0883c6380e0900380ef4457dfa0d367b1bd0deb9