Include name, mode and visibility in typedef's named AST
Summary:Typedefs and constants are currently largely ignored in our IDE code. When working on changing it, one thing that I've found really annoying is that it's possible to declare classes and functions based only on named AST, but typedefs require some additional information from AST that is not in NAST.

This makes writing code awkward, because now I have to thread through the ASTs everywhere where IDE code assumed that NASTs are enough (which is the case for classes and functions).

This diffs adds missing info that makes typedef NAST self-contained for declaration purposes, as evidenced by Decl.typedef_decl function signature.

Reviewed By: int3

Differential Revision: D3205838

fb-gh-sync-id: 4983eb696fb90e2d21a0df3295a47c8926e34247
fbshipit-source-id: 4983eb696fb90e2d21a0df3295a47c8926e34247