Rip out IDE concurrency code
Summary:It was disabled in {D3095831}, the justification for which is also there.

Recent coding shows that it's a pain to keep updating it, so deleting it
completely since it's very unlikely I will pursue this approach anytime soon.

I'm leaving the scheduler, JSON message parsing and "hh ide ." client mode
because those seem self contained and there is a chance I will use them
(without concurrency though).

Reviewed By: int3

Differential Revision: D3210649

fb-gh-sync-id: 467bf8811d764cb3fbcf289c5a0cf0756e6a3e09
fbshipit-source-id: 467bf8811d764cb3fbcf289c5a0cf0756e6a3e09