Don't provide PHP function arguments in stacktrace when core dumping
Summary:Providing arguments to functions in a PHP stacktrace might involve
re-entering the VM (to invoke toString() on objects or resources, for
example). This is very problematic when core dumping as the VM is in an
undefined state at that point.

Since its not even clear what value arguments provide in the PHP stacktrace when
HHVM itself is crashing, its better to just avoid generating them in the first
place.

While we're at it, do some other misc cleanup. Abort immediately if we enter
bt_handler while already crashing. Remove an unused "SegFaulting" global
boolean.

Reviewed By: markw65, edwinsmith

Differential Revision: D3211686

fb-gh-sync-id: 7a9669b0d351fc8d2cf7504eb8ffddc39422c13e
fbshipit-source-id: 7a9669b0d351fc8d2cf7504eb8ffddc39422c13e