Compact when last elem is deleted
Summary:I noticed in some test cases that we had an average of 1.1
probes per hashtable lookups on mixed arrays with size == 0.
It turned out this was because of tombstones, because the array
shrank to zero, but wasn't compacted. Update the compact heuristic
to take care of this edge case.

Reviewed By: markw65

Differential Revision: D3196455

fb-gh-sync-id: 5d71f622fcb51578c32ca23cd8da61d4365c516c
fbshipit-source-id: 5d71f622fcb51578c32ca23cd8da61d4365c516c