Fix potential out-of-bounds array write in dirname_helper
Summary:Most of the callers to dirname_helper copy the string with strndup, but
this causes problems if the string has nulls in it. Namely, we'll only copy up
to the null, which can leave our length out of sync with the actual copied
string size. Namely, if dirname("\x00"), we'll pass an empty string to
dirname_helper, but with a specified length of 1. This can result in an
out-of-bounds array write when dirname_helper attempts to null-terminate the
output.

This could be fixed in a variety of ways, including recalculating the length
with a strlen(), but the best way is to tighten up the interface and make these
kind of errors not possible. First, rename dirname_helper to dirname_impl and
put it in an unnamed namespace. This will keep it from being called except for
the file it lives in. Second, rename dirname_safe to just dirname and remove all
of the overloads except one which takes a String and returns a String. All those
overloads are bug prone and largely aren't used. By only accepting the input as
a String, which can copy it without regards to the contents of the string and
also remove some manual memory management.

Differential Revision: D3186448

fb-gh-sync-id: 83f9a8f0ba70a25f8f25296885251daa50cdbedd
fbshipit-source-id: 83f9a8f0ba70a25f8f25296885251daa50cdbedd