Fix HHI generation on OS X
Summary:
This reverts D2834659 and D3088173, which broke HHI generation on OS X. The options passed are GNU-specific extensions, not available with the BSD tar on OS X (and with no obvious equivalents from a quick skim of the manpage). Unfortunately, although `tar` failed, the shell redirection still generated an empty output file, which caused `make` not to notice the failure.

In the interest of simplicity, and of a quick an unbreak as possible for the Mac users that have complained about this since the 3.13 release a couple of weeks ago, I'm just doing a straight revert here. If the original author (or anyone else) would like to try again, either with more portable `tar` options or with some sort of OS detection/gating, please feel free. In particular I realize build reproducibility is important to the Debian folks. (Though I'm pretty sure there are other reasons our builds aren't reproducible right now; my internal build of `hh_server` still embeds the build time, I think external builds are still the same.)

Ref. https://github.com/hhvm/homebrew-hhvm/issues/32

Reviewed By: paulbiss

Differential Revision: D3219053

fb-gh-sync-id: c37db924f13dc2f980549eb94bdb67e0f7d83872
fbshipit-source-id: c37db924f13dc2f980549eb94bdb67e0f7d83872