Improve local finder error reporting and robustness
Summary:
The local finder is crashing when attempting to add a local to the head of an empty scope chain. Since we always push a scope when entering a function, this means that the crash must be happening because a local is being found by the parser outside of a function.  The visitor pass will then attempt to add it to the top of a non-existing scope chain.

I was unable to find a reproducer for this reported crash, so I'm addressing the problem in three ways.

First, I'm adding a test case that demonstrates that the local finder can get correct results even in a world with an incorrect parse.

Second, the entire service is now wrapped in a try block which adds sufficient context to the failure to enable successful debugging, should we have a regression.

Third, we now put a never-popped scope on the top of every scope chain. In correct code that scope will never have any local added to it, but it might in incorrect code. If my analysis is correct then this will prevent the crash; if incorrect then at least we'll get a better crash dump.

Many thanks to Kasper for assistence analyzing this issue.

Reviewed By: dabek

Differential Revision: D3220625

fb-gh-sync-id: 0a0401e6bb3fa13c63cb65992e6b2f2ec9927cf0
fbshipit-source-id: 0a0401e6bb3fa13c63cb65992e6b2f2ec9927cf0