Check inner types along with guards in retranslation chains
Summary:
Even when a block had a retranslation for a different set of types in a region,
the inner-type checks were only done lazily, which, upon failure, would exit the
region instead of jumping to the next retranslation.  This diff fixes this
problem by eagerly checking the inner types along with the other block guards
for blocks in retranslation chains (i.e. that have a guard failure successor
block).

Reviewed By: paulbiss

Differential Revision: D3223594

fb-gh-sync-id: 14ba8e1444ceef851a041a158f57bf02fc50a9cc
fbshipit-source-id: 14ba8e1444ceef851a041a158f57bf02fc50a9cc