Relax variance restrictions on constraints on class type parameters
Summary:
Covariant and contravariant parameters can safely appear anywhere in "as" and "super" constraints on
class type parameters. Currently an "as" constraint is treated as a negative (contravariant) position,
and a "super" constraint as a positive (covariant) position. This is necessary for constraints on *method*
type parameters but there is no soundness requirement for this check on class type parameters.

(See http://research.microsoft.com/pubs/64042/ecoop06.pdf for a soundness theorem that proves it.)

Reviewed By: dlreeves

Differential Revision: D3235459

fb-gh-sync-id: 483a61d9ddc1201c31a9ef870daaa4a366cc5693
fbshipit-source-id: 483a61d9ddc1201c31a9ef870daaa4a366cc5693