Fix crash in AsioBlockableChain::Unblock
Summary:
If it detects a dependency cycle, it tries to construct an Exception,
which requires that the vmregs are in sync. If they're not, fixupWork
walks the stack until it finds a sync point. There won't actually be a
correct sync point, but generally it will find the one on the call to
join, which is "good enough". However, during warmup we might have
interped the join, and then jitted the async function whose
AsyncRetCtrl caused the unblock, in which case we walk all the way out
and then crash.

This sets up the vmRegs, and marks them clean before calling unblock.

Reviewed By: jano

Differential Revision: D3237102

fb-gh-sync-id: 67a0d540ceceb2eb75646bba0b11d3bcca949ec9
fbshipit-source-id: 67a0d540ceceb2eb75646bba0b11d3bcca949ec9