Trivial tweaks
Summary:
This is preparation for landing support for compressed Phar files.
As discussed in #4263 changes that doesn't affect code logic separated for easier review.
This PR primarily includes manual edits of PhpDoc sections to conform with arguments names from PHP documentation and to make them understandable to IDE. Also some Exceptions classes changed as it is defined in PHP docs.
Closes https://github.com/facebook/hhvm/pull/7019

Reviewed By: Orvid

Differential Revision: D3221580

fb-gh-sync-id: f34bd8dabb7d507db862fe391e7213d90d487963
fbshipit-source-id: f34bd8dabb7d507db862fe391e7213d90d487963