Kill unused and unusable APIs
Summary:
It does not make sense to ever override `Exception::initTrace()`, as `$this->trace`
is private and the only way to set it would be to call `parent::initTrace()`. All
the additional logic of `FooException::initTrace()` could then be moved to
`FooException::__construct()`. Note that if you don't call `parent::initTrace()`,
the exception would be uninitialized and its APIs would fail in `__check_init()`.
It does not make sense to call `parent::initTrace()` later either, because the
while loop would not find the `__init__()` call and clear out the stack
trace completely.

The `{get,set}TraceOptions()` API is not documented and not known to be
overridden. While it could make sense to set up per-subclass trace options,
this API is not particularly nice way to achieve that. Let's mark these
functions as final so they can be used only for their initially intended
functionality.

These APIs were previously build to support hacks in Facebook codebase. These
hacks were removed long time ago, so should be the support code for them.

@override-unit-failures

Reviewed By: markw65

Differential Revision: D3236969

fb-gh-sync-id: c68501366861463c01c72e23376a4df9f6029977
fbshipit-source-id: c68501366861463c01c72e23376a4df9f6029977