Helper functions for outputting multiline positions
Summary:
Pos.t started as "positions of lexer tokens", so every position was on the same line. This has not been true for a long time - we have positions of expressions, which can span multiple lines, and I'm planning to add positions of methods.

But our outputting functions don't support it - they display just 3 numbers, line, start character, and end character. In case of multiline positions the end character will be just garbage out of bonds of current line.

Alternatively, I could represent multiline ranges as two positions - of first and last token, but this is (prohibitively) wasteful from memory perspective. I could also write a function that coverts a multiline position into a pair of 0-length positions representing start and end, but then the end character in the output would be redundant.

Reviewed By: ericlippert

Differential Revision: D3234624

fb-gh-sync-id: 70b1551fb9206da4ee8f740ed03693006ca90c24
fbshipit-source-id: 70b1551fb9206da4ee8f740ed03693006ca90c24