Represent local variables with an abstract type
Summary:
I'm working on refactoring our naming, and this is the first step

Currently, we have an `Ident.t` type that is used for naming local variables. But it's not abstract -- it exposes the fact that it's an integer -- so it's also used for many other things. E.g. expr ids, type variables, and generally any kind of integer identifier. Ideally these would all be separate types so that we can better scope our refactoring. This diff just separates the local variable ids into their own type.

Reviewed By: dlreeves

Differential Revision: D3241193

fb-gh-sync-id: 7529c7fc2fe8f51c1702ed009ace762762ecbc7a
fbshipit-source-id: 7529c7fc2fe8f51c1702ed009ace762762ecbc7a