Add new array-data operations for Hack arrays
Summary:
Add new array-data operations NvTryGetInt, NvTryGetStr, LValIntRef, and
LValStrRef. NvTryGetInt and NvTryGetStr are similar to NvGetInt and NvGetStr
except they're allowed (but not required) to throw if the key isn't
present. LValIntRef and LValStrRef are likewise similar to LValInt and LValStr
except they indicate to the array that the retrieved value will be boxed into a
ref. This is for array types which don't allow refs. In addition, formalize
LValNewRef, which indicates to the array that the new value will be boxed into a
ref.

Try to use the DISPATCH macro in array-data.cpp consistently, using constexpr
aliases in the array implementations to re-use functions.

See: http://hhvm.com/blog/10649/improving-arrays-in-hack
Related to https://github.com/facebook/hhvm/issues/6451
Related to https://github.com/facebook/hhvm/issues/6454

Reviewed By: paulbiss

Differential Revision: D3162677

fb-gh-sync-id: d75e1b554d3c153511177ee2564dfe183c90b888
fbshipit-source-id: d75e1b554d3c153511177ee2564dfe183c90b888