Structures for outline view
Summary:
Right now, outline is just a list of tuples: position, name, description. To know the name of the method, they have to split the name from "class_name::method_name" format, to tell whether method is static they have to distinguish "method" from "static method" description. A lot of things (properties, abstract status, visibility modifiers, extents) are missing. I want to create a new command (--ide-outline?) where we will add this data in more structured way - in this diff I add structures for it (that are limited to the information we currently expose and will be updated later with variants for properties, constants, etc).

It's basically a mini-AST for top level elements. I plan for there to be 1:1 correspondence between it and JSON structure we expose.

Reviewed By: int3

Differential Revision: D3256670

fb-gh-sync-id: 5154c122485feefb4f70543fd0a82d05c32ff21a
fbshipit-source-id: 5154c122485feefb4f70543fd0a82d05c32ff21a