Make PackedArray helpers more generic
Summary:
Make the helper functions used by PackedArray more generic by now longer
assuming the array kind is Packed. The functions will now either copy the kind
of the source parameter, or take the kind as an explicit parameter. In the later
case, the helpers are forced inline, so that the kind parameter will be constant
propagated. This is in preparation of the hack array vec kind, which uses most
of the packed array machinery, but has a different kind.

In addition, move PackedArray::MakePacked (where it isn't used) from
mixed-array.cpp to packed-array.cpp.

There was an implicit invariant that packed arrays all have a non-zero capacity
(thus Grow won't be called on them when empty), which isn't necessarily true for
hack vec arrays, so avoid a memcpy when growing an empty array.

See: http://hhvm.com/blog/10649/improving-arrays-in-hack
Related to https://github.com/facebook/hhvm/issues/6451
Related to https://github.com/facebook/hhvm/issues/6454

Reviewed By: paulbiss

Differential Revision: D3163072

fb-gh-sync-id: a20a2b4a136dcf42754587e836f79b4cd0941e9e
fbshipit-source-id: a20a2b4a136dcf42754587e836f79b4cd0941e9e