Add ToVec array-data function
Summary:
Add a new function to array-data, ToVec. This converts the array to an
equivalent Hack vec array. The conversion is done by discarding the keys and
inserting the values into the new vec array in iteration order. If the array is
already a vec array, itself is returned unchanged. For all array types except
PackedArray, the implementation of the conversion is the same, so defer to a
common helper function to perform it. For a non-vec PackedArray, the conversion
can be done more efficiently since it shares a representation with vec
arrays. The conversion can throw if the source array contains a non-collapsable
reference.

See: http://hhvm.com/blog/10649/improving-arrays-in-hack
Related to https://github.com/facebook/hhvm/issues/6451
Related to https://github.com/facebook/hhvm/issues/6454

Reviewed By: paulbiss

Differential Revision: D3163710

fb-gh-sync-id: 230644c2c932e5849015c36bf15609872f8c1573
fbshipit-source-id: 230644c2c932e5849015c36bf15609872f8c1573