Properties extents in outline
Summary:
This one is more interesting. What is the "property extent?". At first, one might think that in a definition like:

  public ?string $foo = "aaaa";

extent of "foo" is entire line from the start of "public" to the semicolon (excluded).  This is not true though - what here is a single "$foo = "aaaa" is in fact a list, and "public string" applies to all of it's elements. So it could be as well:

  public ?string
     $foo = "aaaa",
     $bar = "cccc",
     $i_have_no_initializer;

which makes it hard to include visibility and type into extent of $bar (since extent is a single continous range).
To make things more complicated, there are two more kind of properties - XHP properties, with syntax

   attributes
      string xhp_prop = "aaa" required,
      int other_xhp_prop;

where each type and "required" applies only to one property and could be part of extents. That IMHO would be weird inconsistency with previous case.

Moreover, there is also the case of implicit properties defined in constructor:

  public function __construct(
     public string $property,
   )

The visibility, type and possible initializer are "per property", but capturing their whole span is annoying from implementation point of view - constructor is just a regular method in AST with special name, so we would have to store additional data for every method argument just to use it in this single case.

The "lowest common denominator" of all those cases is to treat the property extent as span of name + initializer, if present, and this is what that diff implements.

Reviewed By: int3

Differential Revision: D3263186

fbshipit-source-id: b6ae2019f874842ea4af72db6d200433abd8b194