Class constants extents in outline
Summary:
Constants suffer similar problem as properties in D3263186 - the "const <type>" is in fact followed by a list of identifiers, so we can't include them in extents, and extent is just the name + initializer. We could include type and "abstract const" part in the case of abstract constants, but I don't do this for consistency.

Because constants are not complicated by "additional definitions in constructor" like properties were, this diff didn't require adding any new positions to AST - entire extent can be recovered from existing nodes for name and initializer expression.

Reviewed By: alexchow

Differential Revision: D3263191

fbshipit-source-id: 7754fd78e503d77ba80e94c39dea8ce68f2547b2