--ide-outline
Summary:
We want a richer outline view, but --outline is already in use by Nuclide/FBIDE and it's hard to introduce breaking changes, so putting it under a new command.

The correctness of the command is covered in previous diffs by real unit tests, so integration test added here just tests that the command exists, is parsed correctly and outputs something.

Reviewed By: int3

Differential Revision: D3263963

fbshipit-source-id: 7cfb865406ccbc7f2eec7687d7df6c3ab05980c9