Simplify outline structure
Summary:
Changes of the format of outline view after getting feedback from peterhal.

The previous one was complicated and assumed knowing a lot of things about it - you need to know that when kind is "class" then "members" property exists,  that methods can be abstract, but functions not, etc.... This level of structure is not needed by editor.

Changes included:

- every kind is represented by a record of same structure
- every node can have children - right now only classes have non empty children, but if we ever add function arguments here we can represent them as children of function, same for namespaces, etc...
- instead of boolean/variant properties for modifiers varying by the node kind there is a single list of strings to represent them
- renamed "type" to "kind" since it's a word with many meanings

Reviewed By: int3

Differential Revision: D3267756

fbshipit-source-id: 38c67e1d6cfe677b54b483ea04ddf3e2df2901c6