Simplify NativeData scanning
Summary:
We had to define a nativeDataInfoScan<T>() for both cases:
when T::scan() exists, and when it doesn't. The latter uses
conservative scanning, but we never called it; instead we
register a nullptr scan function, then call a generic
conservativeScan() helper function instead. Now we always
use nativeDataInfoScan<T>. Update ZendObject to match this
scheme.

Differential Revision: D3277729

fbshipit-source-id: a59717cd44cc22da94f666bf0dafb7174129519b