Don't barf after ifThen() where the done block is unreachable
Summary:
Currently, ifThen() and friends can, in certain situations, start
emitting code to unreachable blocks.  This occurs if, e.g., a
conditional jump in the branch block is simplified at gen-time to an
unconditional jump, and its target block is unconditionally terminal:

    ifThen(
      env,
      [&] (Block* taken) { gen(env, JmpZero, taken, tmp); },
      [&] { /* some sequence of instructions ending in RetCtrl */ },
    );
    push(env, tmp2);

If `tmp' is a constant, the JmpZero is simplified to a Jmp to `taken'.
Subsequently, since the done block that ifThen() tries to start after
emitting both sides of the branch is unreachable, the out-state of the
taken block is used by default as the in-state for the done block, which
is incoherent.  (In this case, it will have a null fp and sp.)

There are a number of reasonable long-term solutions for this, but for
now, just fix ifThen() and ifElse() when the done block is unreachable.

Reviewed By: swtaarrs

Differential Revision: D3233392

fbshipit-source-id: 12970e6ca50c06cf6713257fade7b7f4e0722c9d