Split IdentifySymbolService.find_symbol_result into SymbolOccurence and SymbolDefinition, reuse FileOutline code to compute definitions
Summary:
Motivation:

IdentifySymbolService.find_symbol_result used to hold information about a particular occurrence of a symbol - something that can be computed just by examining a single file.

I added some members to it that are related to the symbol definition (location and span of definition), and then created very similar structures for FileOutline. This is duplication so it's Bad™.

Merging those structures into one is awkward for two reasons:

- there can be occurrences of symbol that have no definition so everything would be optional for no reason

- there are definition kinds that don't correspond to occurrence kinds - for example, definition is able (and it's a feature requested by Nuclide) to distinguish between class/enum/trait, while it's impossible to tell which one it is just by looking at single file - they look the same and we just return "Class" (which is not really correct, but it was always like that and changing it to something more accurate like "TypeIdentifier" would break editors).

What I did instead is renamed find_symbol_result to SymbolOccurrence, removed definition-related fields from it, extracted the type used in file outline to SymbolDefinition, and made ServerIdentifyFunction return a pair of them,.

That way I am able to re-use function that construct definitions from FileOutline in ServerSymbolDefinition and remove duplication. Unit test show that it benefits few symbol kinds who didn't have span implemented in latter.

There is a lot of renaming, moving and noise in this diff, so I'll try to comment on each file about what is going on.

Reviewed By: dlreeves

Differential Revision: D3280691

fbshipit-source-id: 42a660f9ea26d71e9c7c9c9a73c3b3e46639fb15