better control over brotli window size
Summary:
When creating brotli compressor we need to specify the slidign window size for the
compressor to use. We want to experiment with large windows, but do not want to
waste memory if the blob we are compressing is relatively small.
Notice that Brotli itself makes sure to sanitize the input to be in the
range [2^10, 2^24].

Reviewed By: bmaurer

Differential Revision: D3286762

fbshipit-source-id: a13acb908d4b71c67d32202b48828604d16fcdd6