Profile and try to guard on the types of elements loaded from arrays
Summary:
This diff adds a type profiling mechanism and uses it to improve the handling of
values loaded out of arrays.  When the profiled value is monomorphic, or always
uncounted, or always unboxed, we now check for such property right after loading
the value out of the arrays, which allows better code to be generated by
eliminating or simplifying the unboxing and increfing that are later performed
on the value.

Reviewed By: mxw

Differential Revision: D3257707

fbshipit-source-id: 0fab78176c1647ced4e8cf4e8265d8eb044d4b11