Use code.hot for optimized translations
Summary:
This diff repurposes code.hot to be used for optimized translations.  This can
result to better code locality because: (a) optimized translations, although
tend to be hotter, are generated later on after enough profiling data is
collected; and (b) generating optimized translations in the middle of live
translations can break the natural fall-through among live translations.

This change will also make it easier to track how much code is being JITed in
optimized mode.

Reviewed By: swtaarrs

Differential Revision: D3288786

fbshipit-source-id: 9ada52f8b4bc642f21bd53e6cc57afef2c6a142b