interpreter: Factor out immediate decoding, part 8
Summary:
Refactor instructions that take a variadic number of
immediate arguments: NewStructArray, IterBreak, Switch,
and SSwitch. Each of these takes the non-variadic immediates
as normal parameters, including the count of variadic elements,
then finally a pointer to the first variadic immediate in
the bytecode steam.

Remove the IOP_ARGS macro now that it's dead.

Reviewed By: swtaarrs

Differential Revision: D3140934

fbshipit-source-id: 4cf73a7b631a859d6c292b066f068977ae308fa4