Fix multiline breakpoint does not trigger issue in xdebug
Summary:
There are two problems in previous breakpoint implementation for multiline statement:
```
31 foo(
32    bar(),
33 );
```
1. Breakpoint sets on the first line(31) is not triggering.
2. If you set breakpoints on both line32 and line33, you will hit line33 breakpoint multiple times.

This diff fixes both of them:
1. Instead of comparing source line number for breakpoint, I am changing implementation to only compare offset using m_lineBreakPointFilter which will be set on the first byte code offset of corresponding source line.
2. Breakpoint source line will be bound to end line of the tightest enclose line range.
3. LineToOffsetRangeVecMap is changed to *not* include offset ranges of outer line range for inner lines.

This will work 99.9% of the cases unless there is a new basic block entry-point in the middle of line, but either we do not care about that 0.1% scenario or we can fix it in future.

Btw: the #3 above also seems to fix the source stepping behavior for multiline statement. For example, for "step_over.php" test, it used to be wrong behavior which only step the loop body once even though it executes twice, but it is correct behavior now.

Reviewed By: alexmalyshev

Differential Revision: D3290389

fbshipit-source-id: a6b4efce99beede3da1e66d6b1f51cb8a89a19ea