Fix missing refiness checks in SetWithRefElem
Summary:
SetWithRefElem wasn't properly switching between Define and DefineReffy based on
the reffiness of the rvalue. This was breaking array_map on new hack arrays
which ban references. In the jit it was allowing references to leak into these
types while in the interpreter it was failing even on non reference types.

I also deleted SetWithRefNewElem as it wasn't used anywhere.

Reviewed By: alexmalyshev

Differential Revision: D3295842

fbshipit-source-id: 4bc48ef1ec0a5c6b1c6dafdec2f880da6b55e7b1