Remove unnecessary decode functions, pass arFromInstr instead of origpc
Summary:
Now that immediate decoding takes place in the overloaded
iopWrapper() functions, several decode functions are now dead.

It turns out that all the instructions which took an
"original pc", only needed it as a parameter to arFromInstr(),
to find the callee actrec during the sequence of instructions
preparing for a call. Change them all to take an ActRec*
and move the arFromInstr calls to iopWrapper().

Depends on D3140934

Reviewed By: swtaarrs

Differential Revision: D3141033

fbshipit-source-id: bb5419a981fbde3f64a9146c9db6246ef6e2d616