Cache for frequently read APC entries
Summary:
Add a cache layer for quickly fetching from a subset of APC entries that are rarely updated.

The cache replicates a subset of ConcurrentTableSharedStore's contents. It currently supports singletons and uncounted arrays/strings (relies on treadmill).

A set of key prefixes decides which keys are eligible for the cache. Entries are added on successful slow-case lookup until the cache is full.

Reviewed By: swtaarrs

Differential Revision: D3280890

fbshipit-source-id: ae3e4f67dba5241433c12393f613f5e25d3c6551