Move file outline computation from server to client
Summary:
FileOutline (in it's current form) doesn't require any knowledge from the server, so it doesn't need to be blocked on it's initialization. It''s also pretty time critical, since looking at outline and using it to navigate just opened file is likely to be one of the first thing someone might want to do after launching the editor.

Initially I didn't want to do this because I was afraid of bloating the client and possibly regressing it's startup time, so I considered moving the server-independent stuff to separate binary (similar to hh_format) But it looks like this ship has already sailed and client already depends on everything needed to do this, so moving it is a speed up for free.

Reviewed By: ericlippert

Differential Revision: D3300443

fbshipit-source-id: 7c8c53ebe3df19530149e800b336d3d89815625f