Improve optimizePhis a little
Summary:
After D3287556 we generate quite a few JmpNZ(ExtendsClass)
sequences. ExtendsClass generates a little tree of branches to produce
a boolean result, and then the JmpNZ tests that, and branches on the
result. The final code looks like:

    (14) t6:Bool = ExtendsClass t4:Cls, Cls(X)
        Main:
         0xf2002c2: cmp $0x1c108d18, %rdx
         0xf2002c9: jnz 0xf2002cf
         0xf2002cb: mov $0x1, %bl
         0xf2002cd: jmp 0xf2002ea
         0xf2002cf: cmpl  $0x2, 0xf4(%rdx)
         0xf2002d6: jb 0xf2002e8
         0xf2002d8: cmpq  $0x1c108d18, 0x130(%rdx)
         0xf2002e3: setz %bl
         0xf2002e6: jmp 0xf2002ea
         0xf2002e8: xor %bl, %bl

    (15) JmpZero t6:Bool -> B5<Unlikely>
        Main:
         0xf2002ea: test %bl, %bl
         0xf2002ec: jz 0x16e00a43

We have an optimizePhis step in vasm, but it doesn't quite work for
this case because we end up with a no-op phidef/phijmp block in the
middle which breaks the pattern that optimizePhis is looking for.

I added an additional step to redirect the preds of such blocks to the
target of the phijmp, and we now get:

    (14) t6:Bool = ExtendsClass t4:Cls, Cls(X)
        Main:
         0xf2002c2: cmp $0x1c108d18, %rdx
         0xf2002c9: jz 0xf2002e9
         0xf2002cb: cmpl  $0x2, 0xf4(%rdx)
         0xf2002d2: jb 0x16e00a4f
         0xf2002d8: cmpq  $0x1c108d18, 0x130(%rdx)
         0xf2002e3: jnz 0x16e00a4f

    (15) JmpZero t6:Bool -> B5<Unlikely>

While working on this, I noticed that the optimization was too
conservative with jccs; with setccs its only worth doing the
optimization if the setcc can be pushed back into all preds, but with
jccs its better to push the jcc back into any suitable pred, so I also
fixed that (it doesn't affect this case, unless you remove the first
optimization).

Reviewed By: ricklavoie

Differential Revision: D3295711

fbshipit-source-id: 43eeecd9651c9a734b48a68836cc792fac52ada5