Keep generating PGO translations while there's space in the hot code area
Summary:
We're now using the hot code area for Optimize translations, so it doesn't make
sense to be preventing these translations just because the main area filled up.
This diff keeps the ProfData around until we fill both the hot and main areas or
we have optimized all the functions that were profiled.

Reviewed By: paulbiss, swtaarrs

Differential Revision: D3312953

fbshipit-source-id: 055bf0622f53ae8ec9a366806e1ef728964ef602