Make XboxServer::TaskStartFromNonRequest() sometimes not throw an exception
Summary:
A missing return statement meant this function always threw an exception. I didn't notice this before because the only place it is used (in the HHVM Watchman client extension) captures the exception text and this was never checked.

I will probably alter the HHVM Watchman client extension so it's harder to ignore exceptions like this shortly.

Reviewed By: paulbiss

Differential Revision: D3317778

fbshipit-source-id: c6141a7654248161e8b972d1772636906106a2d3