Starting to get the full fidelity parser reviewed
Summary:
I want to get the full-fidelity parser infrastructure reviewed one piece at a time, to make sure that everyone understands how it is built, and to get feedback on the various parts.

We'll start with the most basic, easy stuff imaginable.  There are three basic kinds of things:

* trivia: whitespace, newlines, comments
* tokens: an individual lexeme, with leading and trailing trivia
* syntax: a structured tree of syntax nodes, where the leaves are tokens.

Each of these has a type which simply describes the kind of thing, without any further information. This will come in handy later.

The syntax kind list is not nearly complete.  The token list is mostly complete; I have not yet added XHP-related tokens.

I've also got a type just to keep track of the different kinds of operators. Grammatical analysis can determine the operator kind.

Each comes with a handy "to_string" for debugging purposes.

Reviewed By: andrewjkennedy

Differential Revision: D3318715

fbshipit-source-id: 93e328584ef891935e651dd34e2a23a490096955