Make jit::ProfData thread-safe
Summary:
This diff adds internal locking/atomicity to all of ProfData's data
members, allowing us to create Profile translations concurrently in multiple
threads. This is done with a mixture of rwlocks, AtomicVectors, and
folly::AtomicHashMaps, depending on the specific needs of each data
member. There's still some disentangling to be done in
MCGenerator::retranslateOpt() before we can concurrently translate Optimize
translations.

The main changes that leaked out of prof-data.{cpp,h} are that
ProfData::curTransID() is gone and ProfData is no longer a member of
Translator. curTransID() has been replaced by ProfData::allocTransID(), which
should be called once at the beginning of the translation process. The caller
is then responsible for threading the returned ID through to any code that
cares about it. I also fixed things up so ProfData and the TransDB use the same
IDs for translations when both PGO and Eval.DumpTC are on.

jit::profData() is the new way to access the singleton ProfData. It reads from
a thread-local copy of the global ProfData* that is only updated once per
request, at the very beginning. This allows us to delete (using Treadmill) the
whole ProfData struct once we're done with it, while guaranteeing that
profData() will always return the same value during the lifetime of any given
request.

Reviewed By: ottoni

Differential Revision: D3053790

fbshipit-source-id: b9c31ed7a672eb276be665f8f31fbc5d3a3d4d59