Keep internal stack frames in Exceptions
Summary:
D2216854 was an incorrect attempt to achieve compatibility with PHP. The real
issue was that $file and $line should always point to the user code, but the
diff author seems to have mistaken it for a non-issue that internal frames are
present in the backtrace and the fix involved off-by-one error, depending on
the last internal frame to still be present.

Let's restore these internal frames (they help debuggability and they are still
present in reentrant code and via every other mechanism that collects stack
traces) and fix the original issue.

If people don't like these internal frames, they should be eliminated in a more
consistent way from all places that invoke stack trace computation logic.

Reviewed By: markw65

Differential Revision: D3321628

fbshipit-source-id: a62ea3d67ac939933760464f3d785bd25c77f805