Do not show implicit properties in outline view
Summary:
Outline view is meant to show outline of the actual file as seen in editor, so it should be a simplified dump of concrete syntax tree. Unfortunately our syntax tree is not a concrete one, and not only does it miss some elements, but also fake nodes are synthesized at parsing level.

This diff ads a reverse logic to FileOutline that filters them out.

Reviewed By: int3

Differential Revision: D3318955

fbshipit-source-id: 074816cf6f6834badb1b98054c5e7bb3c5251160