Bugfixes, more extents, and unit tests
Summary:
Added three new methods source_extent_class_header, source_extent_typeConst, and source_extent_classUse for finding the extents of class headers, typeConsts and classUses.

source_extent_method_ was returning an extend that started with the function keyword. Fixed this to return an extent that includes all of the modifiers on the function.

Also fixed a bug where the result of finding an extent of an empty function would not include the braces.

Added unit tests for these

Added a comment to the top of ast_code_extent explaining how it works and why.

Reviewed By: hermanventer

Differential Revision: D3268338

fbshipit-source-id: f796da88872fe8588a6cefdf6eacbe5c478c2f73