Deprecate Vector::contains
Summary:
`Vector::contains()` was meant to be deprecated for a while now:

https://github.com/facebook/hhvm/blob/HHVM-3.12/hphp/runtime/ext/collections/ext_collections-idl.h#L538

but we never have officially done it. Now that the collections IDL to HNI conversion is done, let's do it.

And we have had people actually using it, confused why it didn't really check for a value existence, as opposed to a key existence.

So raise a deprecation warning for now, and we can officially remove it at a later release. Maybe after the next LTS?

Reviewed By: swtaarrs

Differential Revision: D2828990

fbshipit-source-id: b16a11bde41dec49423bdaf763c37249ecba7181